backend/pyproject.toml
[project]
name = "hvac-assistant-backend"
version = "0.1.0"
description = "AI-powered HVAC technician assistant backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.3.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "httpx>=0.26.0",
    # Google Cloud / Vertex AI
    "google-cloud-documentai>=2.20.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-cloud-storage>=2.14.0",  # For Document AI batch processing
    # Google AI / Gemini (simpler API key approach)
    "google-generativeai>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "api",
    "core",
    "models",
    "services",
    "utils",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
